requirements.txt
psycopg2-binary
requests
pandas
sqlalchemy